-Added getAdress method on motor and changed type to int -Added -1 address verification (-1 means don't initialize)

+++ RPICanDriver/src/Motor_Driver_DSPIC33.cpp
@@ -98,7 +98,7 @@ void Motor::writePos(uint16_t pos){
 
 }
 
-uint8_t Motor::getAdress(){
+int Motor::getAdress(){
 
 	return address;
 }
